+++ src/index.html
@@ -7,9 +7,9 @@
 </head>
 
 <body>
-    <main id="main" class="main">
-        <nav id="nav" class="nav"></nav>
-        <section id="section" class="section"></section>
+    <main>
+        <nav id="header-navigation"></nav>
+        <section class="news-section"></section>
     </main>
     <script src="scripts/index.js"></script>
 </body>
