+++ src/scripts/Content.js
@@ -1,0 +1,9 @@
+export class Content {
+    constructor() {
+        this.isEmpty = function(data) { return data ? data : '' };
+    }
+
+    addContentDocument() {
+        this.element.innerHTML = this.content;
+    }
+}
