+++ src/scripts/index.js
@@ -1,67 +1,12 @@
-import "../style/style.css";
+import "../styles/style.css";
 
 import { config } from "./config";
+import { RequestSources, RequestNews } from "./getUrl";
+import { Data } from "./Data";
+import { Sourses } from "./SourcesContent";
+import { News } from "./NewsContent";
 
 (async () => {
-
-    class RequestSources {
-        getUrl() {
-            return 'https://newsapi.org/v1/sources';
-        }
-    }
-
-    class RequestNews {
-        constructor(target, apiKey) {
-            this.getUrl = () => `https://newsapi.org/v1/articles?source=${target}&apiKey=${apiKey}`;
-        }
-    }
-
-    class Data {
-        constructor(url) {
-            this.url = url.getUrl();
-            this.data = async () => {
-                let response = await fetch(this.url);
-                return response.json();
-            }
-        }
-    }
-
-
-    class Content {
-
-        addContentDocument() {
-            this.element.innerHTML = this.content;
-        }
-    }
-
-    class Sourses extends Content {
-        constructor(data) {
-            super();
-            this.element = document.querySelector('#nav');
-            this.content = data.sources.reduce((acc, item) => {
-                return acc += `<div id="${item.id}" class="source-item">${item.name.length > 10 ? item.name.substring(0,10)+'..' : item.name}
-									<span class="tooltiptext">${item.name}</span>
-				  		 </div>`
-            }, '');
-        }
-    }
-
-    class News extends Content {
-        constructor(data) {
-            super();
-            this.element = document.querySelector('#section');
-            this.content = data.articles.reduce((acc, item) => {
-                return acc += `<a href="${item.url}"><article class="article" ${item.urlToImage ? `style="background-image: url('${item.urlToImage}');"` : ''}>
-									  <h3 class="article-title">${item.title ? item.title : ''}</h3>
-										<p class="article-descr">${item.description ? item.description : ''}
-											<time>${item.publishedAt ? item.publishedAt.substring(0,10) : ''}</time>
-											<span class="author">${item.author ? item.author : ''}</span>
-										</p>
-									 </article></a>`
-            }, '');
-        }
-    }
-
 
     const requestSources = new RequestSources();
     const request = new Data(requestSources);
@@ -71,8 +16,8 @@ import { config } from "./config";
 
     sources.element.addEventListener('click', async (e) => {
         if (e.target.className === 'source-item') {
-            const requestArticles = new RequestNews(e.target.id, config.apiKey);
-            const request = new Data(requestArticles);
+            const requestNews = new RequestNews(e.target.id, config.apiKey);
+            const request = new Data(requestNews);
             const dataNews = await request.data();
             if (dataNews) {
                 const news = new News(dataNews);
